requirements.txt
python-telegram-bot[webhooks]==21.11.1
gspread==6.1.3
oauth2client==4.1.3